Add files via upload
Here are stored the files used for the evaluation of this application. We use a benchmark approach to evaluate the model by applying Jaccard Distance, which checks how many words two sentences share by ignoring the word order. It evaluates this similarity giving it a score from 0 (the same words) to 1 (very different). We compare these results with the ones of LexDiv. We consider a positive evaluation if the change between these two results is 0.1 or lower. This evaluation showed that 91.94% of the predictions were correctly made.